+++ frontend/src/App.js
@@ -1,27 +1,3 @@
-// import React from 'react';
-// import Header from './Header';
-// import HeroSection from './HeroSection';
-// import Categories from './Categories';
-// import Footer from './Footer';
-
-// import './App.css';
-
-// function App() {
-  
-
-//   return (
-//     <div className="App">
-//       <Header />
-//       <HeroSection />
-//       <Categories />
-//       <Footer />
-
-//     </div>
-//   );
-// }
-
-// export default App;
-// App.js
 import React, { useState } from 'react';
 import Header from './Header';
 import HeroSection from './HeroSection';
