+++ frontend/src/Categories.css
